feat: identity + conversations + DCGM telemetry + bundles rename (#324)
* feat: identity (google oauth + admin), conversations, dcgm richer, bundles rename

Big merge covering identity plane, per-user chat history, richer DCGM
telemetry, and the plugins/→bundles/ directory rename we'd been
punting on.

## Identity
- `/auth/google/login` + `/auth/google/callback` (oauth2 authorization
  code flow). `[auth.google]` config block with `client_id`,
  `client_secret_env`, `redirect_uri`, `allowed_domains`, `default_role`.
- `users.google_sub` + `users.avatar_url` columns (migrations 2/4).
- `sessions::upsert_user_from_google` (by-sub then by-email link) and
  `sessions::create_session_for_user` (password-less trusted path).
- Server-side state cache alongside the cookie so OAuth survives strict
  third-party-cookie policies / multi-tab sign-in.
- `/web/login` page with Sign-in-with-Google button + password fallback.
- Shell gate redirects unauthenticated visitors to `/web/login`.
- `whoami` now returns `email`, `display_name`, `role`, `avatar_url`.
- Chat + StatusStrip render the Google profile picture and display name;
  the chat handler injects `<gadgetron_user>` into shared_context so
  Penny can greet the caller by name.
- Admin page `/web/admin` wired to the existing users CRUD endpoints
  (list, add, delete). Top-bar Admin/User view toggle hides the admin
  rail entry in User mode. CLI `key create --scope "A,B"` now splits
  correctly.

## Conversations (ChatGPT-style left-rail)
- `conversations` table + soft-delete + `(user_id, updated_at)` index.
- Backend: `GET/DELETE/PATCH /workbench/conversations[/{id}]` +
  `GET /workbench/conversations/{id}/messages` reading the Claude Code
  jsonl and returning structured blocks (text / reasoning / tool_use /
  tool_result).
- Chat handler upserts a row per turn via `X-Gadgetron-Conversation-Id`
  and treats the id as the Claude `--session-id` so the jsonl path is
  predictable.
- Frontend: ConversationsPane in the left rail (new-chat button,
  click-to-switch with reload, hover delete, 5 s refresh + on-stream
  refresh). Composer persists a per-conversation draft across reloads.
- PastMessages renders the history above the live thread with markdown
  and a date divider; tool_use / tool_result blocks are filtered out
  of the transcript view per request.

## DCGM telemetry
- Added fields: `mem_temp_c`, `ecc_dbe_total`, `xid_last`,
  `throttle_reasons` + decoded label.
- Header-based dmon parser — robust to DCGM 3.x column reorders that
  broke the old positional parser. Field list now
  `203,204,150,140,155,160,112,230,310,319`.
- Per-host DCGM_HEALTHY cache + `timeout 1s` wrapper: broken
  `nv-hostengine` no longer stalls polling for 5 s every tick.
- Bootstrap install script verifies nvidia-dcgm actually reaches
  `is-active` and passes `dcgmi dmon` probe before reporting success.
- Card: GpuHealthBadges (ECC / XID / throttle with shorthand cause).
- Drawer: `GPU memory temp`, `GPU throttle` chart groups with decoded
  y-axis.

## Server-monitor
- Server-side background poller (`bundles/server-monitor/src/poller.rs`)
  — 1 Hz JoinSet across hosts, shares metrics-writer channel with the
  gadget path. Continuous `host_metrics` even with no UI open.
- `server.systemctl` + `server.journal` gadgets. NOPASSWD sudoers now
  covers `/usr/bin/systemctl` + `/usr/bin/journalctl`.
- Default tenant for new hosts is the demo tenant id; prevents the
  "nil tenant_id → empty drawer graphs" regression.
- Timing instrumentation on `call_stats`.

## plugins/ → bundles/ rename
- `plugins/plugin-server-monitor` → `bundles/server-monitor`
  (merged with the pre-existing `bundle.toml`).
- `plugins/plugin-document-formats` → `bundles/document-formats`.
- Workspace `Cargo.toml` + dependent crate path rewrites.
- `plugins/` directory removed.

## Gateway / penny / misc
- `samples::bigint` cast in server_metrics.rs — fixes 400 on tiers
  above `raw` where `SUM(bigint)` widens to numeric.
- `serve_stdio` dispatches each tools/call on a spawned task with a
  single writer task draining through an mpsc channel — parallel tool
  calls no longer trip the MCP client timeout.
- `SessionStore::get_or_create` reuses the conversation UUID as the
  Claude session UUID when the id parses cleanly.
- Penny persona gained a "Voice" section — teenage-detective-niece
  speech style per request.

## UI polish
- Chat column widened to `min(1400px,92vw)`.
- Bottom typing indicator pulses while Penny composes.
- StatusStrip pruned (plugs strip + gateway-healthy label removed;
  health only appears when degraded/blocked).
- Wiki page list rendered as a collapsible tree with persisted
  expansion state.
- Host-detail drawer: per-series legend toggle; per-host localStorage
  for range + enabled groups + hidden series.
- Host card: age label stabilized ("live" / "updated Ns ago") with
  no pulse animation; drawer Refresh button stops swapping label on
  auto-refresh.
- Google-hosted avatar images sent with `referrerPolicy="no-referrer"`
  so Google's referer-gated images actually load.
- Version badge (bottom-right) inlines `CARGO_PKG_VERSION` at build
  time via `NEXT_PUBLIC_GADGETRON_VERSION`.

## Migrations
- 20260421000002 users_google_sub
- 20260421000003 conversations
- 20260421000004 users_avatar_url


* ci: rustfmt + clippy + test struct fixes, security advisory ignore

- `cargo fmt` across the diff (diverged formatting on oauth / sessions
  / handlers / gadget_server / workbench / server / bootstrap).
- Bench + 4 integration tests + testing harness now fill the new
  `google_oauth: None` AppState field.
- `GpuStats` test fixture in metrics.rs fills the DCGM-only fields
  (mem_temp, ecc_dbe, xid, throttle) as `None`.
- clippy -D warnings fixes in plugin-server-monitor + gateway:
    * `parse_ipmi_sensor_psu_watts` marked `#[allow(dead_code)]`
      (kept for the `[server_monitor].slow_ipmi = true` opt-in path).
    * `run_bootstrap::private_key_path` takes `&Path` (was `&PathBuf`).
    * CPU util delta uses `saturating_sub`.
    * `server_metrics::fetch_points` marked
      `#[allow(clippy::too_many_arguments)]`.
    * `ok_or_else(|| …TenantNotFound)` → `ok_or(...)` (3 sites).
    * `strip_penny_context_prefix` rewritten as `while let Some(a)` loop.
- cargo-deny: ignore RUSTSEC-2026-0104 (rustls-webpki CRL parsing
  panic) — CRL parsing is not on our verifier path; patched version
  isn't yet in the dep tree.


* ci: switch test Postgres image to timescale/timescaledb-ha:pg16

Migration 20260421000001 does `CREATE EXTENSION timescaledb`, so CI
needs a Postgres image that bundles the extension. `pgvector/pgvector:
pg16` blew up the Test job at migration run. `timescale/timescaledb-
ha:pg16` ships both TimescaleDB and pgvector (the knowledge layer
needs the latter), so a single image covers both migrations.


---------

+++ crates/gadgetron-gateway/src/web/server_metrics.rs
@@ -200,6 +200,7 @@ pub async fn list_server_metrics(
     }))
 }
 
+#[allow(clippy::too_many_arguments)]
 async fn fetch_points(
     pool: &PgPool,
     tenant_id: Uuid,
@@ -253,8 +254,14 @@ async fn fetch_points(
         }
         _ => {
             // Continuous-aggregate views: use the `bucket` column.
+            // `samples` is `bigint` in the 5s rollup but becomes `numeric`
+            // from the 1m tier upward — Postgres widens `SUM(bigint)` to
+            // `numeric` to avoid overflow. sqlx can't decode `numeric`
+            // into `i64` without the `bigdecimal` feature, so we cast
+            // back to `bigint` in SQL. Overflow is not a real concern
+            // here (samples per hour ≪ 2^63).
             let sql = format!(
-                "SELECT bucket AS ts, avg, min, max, samples \
+                "SELECT bucket AS ts, avg, min, max, samples::bigint \
                  FROM {rel} \
                  WHERE tenant_id = $1 AND host_id = $2 AND metric = $3 \
                    AND bucket >= $4 AND bucket < $5 \
